Add custom panelization script
The KiKit CLI cannot create a grid with boards rotated differently, so I created
a custom script using KiKit to create the panels. This places two rows of the
board with the SMA connector facing outward, so that they can be soldered on
by the fab without breaking the boards of out the panel.